+++ CHANGES.rst
@@ -9,6 +9,11 @@ Changelog for zest.releaser
   Fixes `issue 311 <https://github.com/zestsoftware/zest.releaser/issues/311>`_.
   [maurits]
 
+- Better support for ``zestreleaser.towncrier`` (and similar extensions):
+  the update_history setting is now also respected by the ``bumpversion`` command.
+  Fixes `issue 310 <https://github.com/zestsoftware/zest.releaser/issues/310>`_.
+  [maurits]
+
 
 6.16.0 (2019-01-17)
 -------------------