+++ CHANGES.rst
@@ -1,7 +1,7 @@
 Changelog for zest.releaser
 ===========================
 
-6.16.1 (unreleased)
+6.17.0 (unreleased)
 -------------------
 
 - Refuse to edit history header when it looks to be from an already released version.